+++ src/types/types.rs
@@ -221,7 +221,7 @@ mod tests {
             PlankType::infer_type(r#"{"name": "me", "age": 10}"#),
             PlankType::Struct(vec![
                 PlankField::new("name", PlankType::Str),
-                PlankField::new("age", PlankType::Int64)
+                PlankField::new("age", PlankType::Int32)
             ])
         )
     }
